Improve hash performance in random OT extension
We introduced a regression when switching to generic curves.
The issue was that we were basically doing double the hashes
by using the Scalar::random method.

Ideally, we would do a wide reduction, and thus need to only read one
message.
Unfortunately, this doesn't seem possible generically, but we can do
rejection sampling instead.